opt/facedetect/requirements.txt
dlib==19.23.1
numpy==1.22.3
Pillow==9.1.0
psycopg2==2.9.3
